requirements.txt
# Django Core
Django>=5.0,<6.0
django-environ>=0.11.2

# Database
psycopg2-binary>=2.9.9

# Image Processing
Pillow>=10.2.0

# Data Analytics & Visualization
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Development
django-debug-toolbar>=4.2.0

# Production (optional)
gunicorn>=21.2.0
whitenoise>=6.6.0
